feat: outsource pairing logic to seperate file
