+++ src/actions/game.ts
@@ -11,93 +11,7 @@ import { GameResult } from "@/db/types/game";
 import { eq, InferInsertModel } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
 import invariant from "tiny-invariant";
-
-const weekdayToIndex = {
-  monday: 1,
-  tuesday: 2,
-  wednesday: 3,
-  thursday: 4,
-  friday: 5,
-  saturday: 6,
-  sunday: 0,
-} as const;
-
-function addDays(d: Date, days: number) {
-  const out = new Date(d);
-  out.setDate(out.getDate() + days);
-  return out;
-}
-
-/** first date ≥ `start` that falls on the wanted `weekday`              */
-function firstMatchDate(start: Date, weekday: keyof typeof weekdayToIndex) {
-  const wanted = weekdayToIndex[weekday];
-  const d = new Date(start);
-  while (d.getDay() !== wanted) d.setDate(d.getDate() + 1);
-  return d;
-}
-
-export function bergerFide(n: number): Array<Array<[number, number]>> {
-  if (n % 2) {
-    throw new Error("n must be even (add a bye for odd n)");
-  }
-
-  const half = Math.floor(n / 2); // boards per round
-  const circle = n - 1; // players 1 … n-1 live on the circle
-  const rounds: Array<Array<[number, number]>> = [];
-
-  for (let r = 1; r < n; r++) {
-    // 1-based round counter
-    // ---------- board 1 (always contains player n) ----------
-    let board: Array<[number, number]>;
-    let startWhite: number;
-
-    if (r & 1) {
-      // odd round
-      const opp = Math.floor((r + 1) / 2); // 1,2,3,4,…
-      board = [[opp, n]]; // n is Black
-      startWhite = (opp % circle) + 1; // first White on the circle
-    } else {
-      // even round
-      const opp = half + Math.floor(r / 2); // n/2+1, n/2+2,…
-      board = [[n, opp]]; // n is White
-      startWhite = ((half + Math.floor(r / 2)) % circle) + 1;
-      if (startWhite === n) {
-        // can never be player n
-        startWhite = 1;
-      }
-    }
-
-    // ---------- remaining boards ----------
-    let wIdx = startWhite - 1; // 0-based index on circle
-    let bIdx = (wIdx - 2 + circle) % circle; // "two steps behind" (with proper modulo)
-
-    for (let i = 0; i < half - 1; i++) {
-      const white = wIdx + 1;
-      const black = bIdx + 1;
-      board.push([white, black]);
-      wIdx = (wIdx + 1) % circle; // walk clockwise
-      bIdx = (bIdx - 1 + circle) % circle; // walk anti-clockwise (with proper modulo)
-    }
-
-    rounds.push(board);
-  }
-
-  return rounds;
-}
-
-/** Berger algorithm for round-robin tournaments */
-function roundRobinPairs(n: number): Array<Array<[number, number]>> {
-  const isOdd = n % 2 === 1;
-
-  if (isOdd) {
-    const extended = bergerFide(n + 1);
-    return extended.map((round) =>
-      round.filter(([white, black]) => white <= n && black <= n),
-    );
-  } else {
-    return bergerFide(n);
-  }
-}
+import { addDays, firstMatchDate, roundRobinPairs } from "@/lib/pairing-utils";
 
 export async function removeScheduledGames(tournamentId: number) {
   await db.delete(game).where(eq(game.tournamentId, tournamentId));
